# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Daily Scheduler",
    "description": "AI-powered daily news & trading report system",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reports": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Reports",
        "description": "List reports with pagination.",
        "operationId": "list_reports_api_reports_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Per Page"
            }
          },
          {
            "name": "report_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Report Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportOut"
                  },
                  "title": "Response List Reports Api Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reports/{report_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report",
        "description": "Get a specific report by ID.",
        "operationId": "get_report_api_reports__report_id__get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reports/{report_id}/html": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Html",
        "description": "Get raw HTML content of a report, optionally in a translated language.",
        "operationId": "get_report_html_api_reports__report_id__html_get",
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Language code; serves a translation when set",
              "default": "",
              "title": "Lang"
            },
            "description": "Language code; serves a translation when set"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "generation_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Time S"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "ko"
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/ReportTranslationOut"
            },
            "type": "array",
            "title": "Translations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "report_type",
          "summary",
          "generation_time_s",
          "created_at",
          "html_content"
        ],
        "title": "ReportDetailOut",
        "description": "Report detail with HTML content + any translated renderings.\n\n``language`` is the primary (generated) language; ``translations`` carries\nthe additional language renderings that power the dashboard language toggle."
      },
      "ReportOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_date": {
            "type": "string",
            "format": "date",
            "title": "Report Date"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "generation_time_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Time S"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "report_date",
          "report_type",
          "summary",
          "generation_time_s",
          "created_at"
        ],
        "title": "ReportOut",
        "description": "Report list item."
      },
      "ReportTranslationOut": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content"
          }
        },
        "type": "object",
        "required": [
          "language",
          "html_content"
        ],
        "title": "ReportTranslationOut",
        "description": "A translated rendering of a report."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}